Refactor main page file structure to follow preferred pattern
